backend/settings_app/serializers.py
from rest_framework import serializers
from .models import Store


class StoreSerializer(serializers.ModelSerializer):
    class Meta:
        model = Store
        fields = [
            'id', 'store_name', 'address', 'contact_number',
            'tax_rate', 'currency', 'receipt_header', 'receipt_footer',
            'created_at', 'updated_at',
        ]
        read_only_fields = ['id', 'created_at', 'updated_at']
